ref(auto_source_config): Move task module into central location
This change brings the task that schedules automatic code mappings into the central location for the code.
